requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
supabase==2.3.0
dash==2.14.2
plotly==5.18.0
Flask-WTF==1.2.1
gunicorn==21.2.0
python-jose==3.3.0
requests==2.31.0
Werkzeug==3.0.1
Jinja2==3.1.2 
pandas